Fix build problems under gcc 4.4 and VS 2008
git-svn-id: https://scantailor.svn.sourceforge.net/svnroot/scantailor/trunk@367 80b70061-164d-0410-b7b7-e274ec932871

+++ imageproc/Morphology.cpp
@@ -309,7 +309,7 @@ void spreadInDirection(
 		return;
 	}
 	
-	int const first_phase_steps = (int)sqrt(num_steps);
+	int const first_phase_steps = (int)sqrt((double)num_steps);
 	
 	BinaryImage tmp(tmp_images.retrieveOrCreate(tmp_image_size));
